requirements.txt
psycopg2==2.9.6
flask==2.3.2
waitress==2.1.2
python-dotenv==1.0.0